+++ src/C0_data_processing.py
@@ -67,11 +67,11 @@ def all(dict_values):
     # C1.check_input_values(dict_values)
     # todo Check, whether files (demand, generation) are existing
 
-    # check electricity price >= feed-in tariff todo: can be integrated into check_input_values() later
-    C1.check_feedin_tariff(dict_values=dict_values)
-
     # Adds costs to each asset and sub-asset
     process_all_assets(dict_values)
+
+    # check electricity price >= feed-in tariff todo: can be integrated into check_input_values() later
+    C1.check_feedin_tariff(dict_values=dict_values)
 
     # Perform basic (limited) check for moduel completeness
     C1.check_for_sufficient_assets_on_busses(dict_values)
@@ -1052,6 +1052,17 @@ def define_source(
                 }
             )
         if item == "price":
+            if FILENAME in kwargs[item] and HEADER in kwargs[item]:
+                kwargs[item].update(
+                    {
+                        VALUE: get_timeseries_multiple_flows(
+                            dict_values[SIMULATION_SETTINGS],
+                            default_source_dict,
+                            kwargs[item][FILENAME],
+                            kwargs[item][HEADER],
+                        )
+                    }
+                )
             determine_dispatch_price(dict_values, kwargs[item], default_source_dict)
 
     dict_values[ENERGY_PRODUCTION].update({asset_key: default_source_dict})